move deb version to new file
